refactor(router): share media speech facts
